requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
databases==0.7.0
docopt==0.6.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
pipreqs==0.4.13
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.23.0
watchfiles==0.19.0
websockets==11.0.3
yarg==0.1.9
